Extract source-map files from inlined source to separate files
