+++ package.json
@@ -39,10 +39,11 @@
       "grunt-contrib-cssmin": "~0.7.0",
       "grunt-contrib-imagemin": "~0.9.4",
       "grunt-contrib-jshint": "~0.8.0",
-      "grunt-contrib-uglify": "~0.3.0",
+      "grunt-contrib-uglify": "~0.10.0",
       "grunt-replace": "0.6.2",
       "grunt-contrib-watch": "git://github.com/gruntjs/grunt-contrib-watch.git#b884948805940c663b1cbb91a3c28ba8afdebf78",
       "grunt-contrib-jasmine": "~0.8.1",
+      "grunt-exorcise": "~2.1.0",
       "grunt-prompt": "~1.3.0",
       "leaflet": "0.7.3",
       "mustache": "1.1.0",
